oss-report-agent/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
PyYAML==6.0.3
pydantic==2.11.7
